requirements.txt
pyyaml
watchdog
pypinyin